Solucion al linter
