+++ .eslintignore
@@ -1,1 +1,2 @@
 src/data
+src/template/graficos.js
